requirements.txt
PyPDF2==3.0.1
python-docx==1.1.0
langchain==0.1.0
langchain-community
langchain-huggingface
langchain-chroma
sentence-transformers==2.2.2
chromadb==0.4.22
transformers
torch
accelerate